Move hierarchy flattening into Veryl frontend
