+++ crates/celox/src/logic_tree.rs
@@ -1,6 +1,5 @@
 mod comb;
 pub use celox_slt::const_inline;
-pub use comb::SLTLoopBound;
 pub use comb::SLTNode;
 pub use comb::{LogicPath, LogicPathTarget};
-pub use comb::{NodeId, SLTForFoldGroupState, SLTNodeArena, SLTNodeFactsError, get_width};
+pub use comb::{NodeId, SLTNodeArena, SLTNodeFactsError};
